check underlying isue before deleting in event handler

+++ app/Listeners/IssueDeletedListener.php
@@ -26,6 +26,8 @@ class IssueDeletedListener
      */
     public function handle(IssueDeleted $event)
     {
-        $event->issue->issue->delete();
+        if($event->issue->issue) {
+            $event->issue->issue->delete();
+        }
     }
 }
